// src/modules/MyBio/models/Kin.ts
export interface Kin {
  id: string;
  firstName: string;
  lastName: string;
  sex: string;
  relationship: string;
  isEmergency: boolean;
  address: string;
  state?: string;
  country?: string;
  mobile?: string;
  home?: string;
  email?: string;
  avatar?: string;
  percentage?: string;
  dateOfBirth?: string;
}